% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{add_route}
\alias{add_route}
\alias{add_route.default}
\alias{add_route.router}
\title{add_route: add new route to a router}
\usage{
add_route(router, method, endpoint, fun)

\method{add_route}{default}(router, method, endpoint, fun)

\method{add_route}{router}(router, method, endpoint, fun)
}
\arguments{
\item{router}{router class}

\item{method}{http verb such "GET", "POST", "PUT" or "DELETE"}

\item{endpoint}{string e.g. "/add/", slashes are important!!}

\item{fun}{function}
}
\value{
router

\code{Object}

\code{Object}
}
\description{
A description of add_route
}
\details{
A details of add_route
}
\examples{
r <- router()
r <- add_route(r,
    method = "GET",
    endpoint = "/add/",
    fun = function(a, b) {
        return(as.numeric(a) + as.numeric(b))
    }
)

}
