% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_custo.R
\name{gerar_custo}
\alias{gerar_custo}
\title{Produz os coeficientes do vetor custo do modelo de otimização linear de roteamento de veículos}
\usage{
gerar_custo(i_limit, j_limit, k_limit, vetor_custos = NULL)
}
\arguments{
\item{i_limit}{Número inteiro que representa a quantidade de pontos (cidades) de partida no modelo, incluindo a origem.}

\item{j_limit}{Número inteiro que representa a quantidade de pontos (cidades) de chegada no modelo, incluindo a origem como destino final.}

\item{k_limit}{Número inteiro que representa a quantidade de veículos que partem da origem.}

\item{vetor_custos}{Vetor de números reais representando a distância entre o trecho (ij) no veículo k, deve seguir a ordem estabelecida na sequencia de variáveis do modelo. Caso não seja fornecido, a função gera valores aleatórios (runif) e simétricos, ou seja distância entre i e j é a mesma entre j e i.}
}
\value{
Dataframe contendo os coeficientes da função objetivo do modelo, ou seja, as distâncias entre cada trecho (ij) no veículo k.
}
\description{
Produz um vetor contendo os custos (distância entre dois nós) para o modelo matemático
}
\examples{
# Exemplo de uso da função sem vetor_custos (usando runif para gerar custos)
resultado1 <- gerar_custo(5, 5, 3)
#print(resultado1)
# Exemplo de uso da função com vetor_custos fornecido pelo usuário
vetor_custos_usuario <- c(10, 20, 30, 40, 50, 60, 70, 80, 90, 100)
#resultado2 <- gerar_custo(7, 10, 3, vetor_custos_usuario)
#print(resultado2)
# Exemplo de uso da função com vetor_custos fornecido pelo usuário
vetor_custos_usuario <- c(10, 20, 30, 40, 50, 60)
resultado3 <- gerar_custo(2, 2, 3, vetor_custos_usuario)
#print(resultado3)

n=5
k=1

vetor_custos=c(2.8,2,100,100,
+ 2.8,2.4,4.4,5.2,
+ 2,2.1,4.2,5,
+ 100,3.6,4.1,0.65,
+ 0,0,0,0)
custo <- gerar_custo(n,n,k,vetor_custos)
print(custo)
custo2 <- gerar_custo(-1, 5, 3)
#print(custo2)
}
\author{
Luciane Ferreira Alcoforado
}
