% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_restricoes_saida_origem.R
\name{gerar_restricoes_saida_origem}
\alias{gerar_restricoes_saida_origem}
\title{Gera restrições de saída da origem}
\usage{
gerar_restricoes_saida_origem(i_limit, j_limit, k_limit, n_restricao)
}
\arguments{
\item{i_limit}{número de nós do problema que representam a origem num trecho da rota}

\item{j_limit}{número de nós do problema que representam o destino num trecho da rota}

\item{k_limit}{número de veículos do problema}

\item{n_restricao}{número de controle da restrição, por padrão terá ser valor igual a 1, podendo ser renumerado de acordo com a modelagem adotada.}
}
\value{
retorna um dataframe contendo as restrições de saída da origem.
}
\description{
Gera as restrições de saída da origem do modelo, ou seja, todo veículo deve sair do ponto 1 (origem).
}
\examples{
# Exemplo de uso da função
restricoes_saida <- gerar_restricoes_saida_origem(5,5,3,1)
print(restricoes_saida)
restricoes_saida1 <- gerar_restricoes_saida_origem(5,5,3,11)
print(restricoes_retorno1)

}
\author{
Luciane Ferreira Alcoforado
}
