% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_restricoes_subrota.R
\name{gerar_restricoes_subrota}
\alias{gerar_restricoes_subrota}
\title{Gera restrições de subrota}
\usage{
gerar_restricoes_subrota(i_limit, j_limit, k_limit, n_restricao, n = 2)
}
\arguments{
\item{i_limit}{número de nós do problema que representam a origem num trecho da rota}

\item{j_limit}{número de nós do problema que representam o destino num trecho da rota}

\item{k_limit}{número de veículos do problema}

\item{n_restricao}{número de controle da restrição, por padrão terá ser valor igual a 1, podendo ser renumerada de acordo com a modelagem adotada.}

\item{n}{número de pontos na rota, ou seja, sub rota de tamanho 1 terá trecho com dois pontos, rota de tamanho 2 terá trecho com 3 pontos, ...Por padrão n=2, ou seja, são geradas restrições de sub rota de tamanho 1.}
}
\value{
retorna um dataframe com as restrições de sub rota de tamanho n-1
}
\description{
Gera restrições de subrota de tamanho n-1, em que n representa o número de pontos na sub rota e deve variar de 2 até i_limit. Por padrão n = 2.
}
\examples{
# Exemplo de uso da função para sub rotas de tamanho 2
restricoes_subrota2 <- gerar_restricoes_subrota(5,5,3,1,3)
print(restricoes_subrota2)
# Exemplo de uso da função para sub rotas de tamanho 1
restricoes_subrota1 <- gerar_restricoes_subrota(5,5,3,1)
print(restricoes_subrota1)
# Exemplo de uso da função para sub rotas de tamanho 3
restricoes_subrota3 <- gerar_restricoes_subrota(5,5,3,1,4)
print(restricoes_subrota3)
restricoes_subrota4 <- gerar_restricoes_subrota(6,5,1,1,4)
print(restricoes_subrota4)
}
