% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_restricoes_unicidade.R
\name{gerar_restricoes_unicidade}
\alias{gerar_restricoes_unicidade}
\title{Gera restrição de unicidade}
\usage{
gerar_restricoes_unicidade(i_limit, j_limit, k_limit, n_restricao)
}
\arguments{
\item{i_limit}{número de nós do problema que representam a origem num trecho da rota}

\item{j_limit}{número de nós do problema que representam o destino num trecho da rota}

\item{k_limit}{número de veículos do problema}

\item{n_restricao}{número de controle da restrição, por padrão terá ser valor igual a 1, podendo ser renumerada de acordo com a modelagem adotada.}
}
\value{
retorna um dataframe com as restrições de unicidade de visita de um dos veículos
}
\description{
Cada cliente (nó) recebe visita de apenas um dos veículos
}
\examples{
# Exemplo de uso da função
restricoes_unicidade <- gerar_restricoes_unicidade(5,5,3,7)
print(restricoes_unicidade)
restricoes_unicidade <- gerar_restricoes_unicidade(5,5,1,1)
print(restricoes_unicidade)
}
