% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_variaveis.R
\name{gerar_variaveis}
\alias{gerar_variaveis}
\title{Gera nomes das variáveis}
\usage{
gerar_variaveis(i_limit, j_limit, k_limit = 1)
}
\arguments{
\item{i_limit}{Número inteiro que representa a quantidade de pontos (cidades) de partida no modelo, incluindo a origem.}

\item{j_limit}{Número inteiro que representa a quantidade de pontos (cidades) de chegada no modelo, incluindo a origem como destino final.}

\item{k_limit}{Número inteiro que representa a quantidade de veículos que partem da origem.}
}
\value{
vetor caracter contendo os nomes das variáveis do modelo
}
\description{
Gera um vetor contendo os nomes das variáveis x_{ijk}. Na modelagem x_{ijk} é uma variável binária igual a 1 se o trecho (ij) for incluído na rota com o veículo k; e igual a zero se o trecho não for incluido na rota com o veículo k.
}
\examples{
gerar_variaveis(3,3)
gerar_variaveis(4,4,2)

}
\author{
Luciane Ferreira Alcoforado
}
