% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordenar_rota.R
\name{ordenar_rota}
\alias{ordenar_rota}
\title{Função para ordenar os pontos de partida}
\usage{
ordenar_rota(rotas_tipo)
}
\arguments{
\item{rotas_tipo}{Dataframe com duas colunas: uma contendo a rota e outra contendo o tipo de veículo}
}
\value{
Dataframe com o tipo de veículo, a rota, a origem e o destino de cada rota
}
\description{
Organiza a solução do modelo por tipo de veículo e rota, explicitando o início e fim de cada rota.
}
