% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizar_rota.R
\name{organizar_rota}
\alias{organizar_rota}
\title{Organiza a rota para cada veículo}
\usage{
organizar_rota(vetor)
}
\arguments{
\item{vetor}{nomes das variáveis de valor 1 na solução do modelo (solução obtida utilizando por exemplo o pacote`lpSolve`)}
}
\value{
Dataframe contendo os índices ij de partida e chegada e o código do veículo correspondente à rota.
}
\description{
Organizar a rota para cada veiculo com base em uma solução
}
\examples{

# Exemplo de uso da função
vetor <- c("x121", "x122", "x125", "x211", "x212", "x213", "x311", "x312", "x313", "x135")
rotas_por_tipo <- organizar_rota(vetor)

}
\author{
Luciane Ferreira Alcoforado
}
